ms2-parking-spots/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.13
pydantic-settings==2.6.0
redis==5.2.0
kafka-python==2.0.2
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==6.33.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
